requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
sqlparse
